refactor: session dependency filename
