+++ app/auth/api/endpoints.py
@@ -1,20 +1,26 @@
 from typing import Annotated
-from fastapi import APIRouter, HTTPException, Header, status, Response, Request
 
-from app.auth.use_cases.reset_password_use_case import ResetPasswordUseCase
-from app.common.api.dependencies.get_session import SessionDependency
-from app.auth.schemas.auth_schema import PasswordResetRequest, UserLogin
+from fastapi import APIRouter
+from fastapi import Header
+from fastapi import HTTPException
+from fastapi import Request
+from fastapi import Response
+from fastapi import status
+from slowapi import Limiter
+from slowapi.util import get_remote_address
+
+from app.auth.exceptions.invalid_credentials_exception import (
+    InvalidCredentialsException,
+)
+from app.auth.schemas.auth_schema import PasswordResetRequest
+from app.auth.schemas.auth_schema import UserLogin
 from app.auth.use_cases.auth_user_use_case import AuthUserUseCase
+from app.auth.use_cases.reset_password_use_case import ResetPasswordUseCase
+from app.common.api.dependencies.session_dependency import SessionDependency
 from app.common.exceptions.model_not_found_exception import (
     ModelNotFoundException,
 )
 from app.core.config import get_settings
-from app.auth.exceptions.invalid_credentials_exception import (
-    InvalidCredentialsException,
-)
-from slowapi import Limiter
-from slowapi.util import get_remote_address
-
 
 router = APIRouter()
 settings = get_settings()
